requirements.txt
python==3.7
asgiref==3.2.7
certifi==2020.6.20
Django==3.0.3
django-ckeditor==5.9.0
django-js-asset==1.2.2
gunicorn==20.0.4
imgaug==0.4.0
matplotlib==3.0.3
msgpack-numpy==0.4.4.3
Pillow==7.2.0
pytz==2020.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Shapely==1.7.0
sqlparse==0.3.1
tabulate==0.8.7
tensorpack==0.10
wincertstore==0.2